Fix warcasket foundry exception related to inspect string
In 1.5, `CompReportWorkSpeed` received a `CompProperties_ReportWorkSpeed` class that corresponds to it. It always assumes that the class is non-null, which causes issues here as warcasket foundry did not define the props class, since it was not needed until now.

This should fix it by defining the props class instead of just providing the `compClass`.